+++ SymbolicLab2/Reader/Models/JsonTerm.cs
@@ -5,7 +5,7 @@
         public Operation Operation { get; set; }
         public JsonTerm FirstParam { get; set; }
         public JsonTerm SecondParam { get; set; }
-        public int? NumberValue { get; set; }
+        public double? NumberValue { get; set; }
         public char? SymbolicValue { get; set; }
     }
 }
